requirements.txt
click==7.1.2
Jinja2==2.11.2
Markdown==3.3.3
MarkupSafe==1.1.1
PyYAML==5.3.1
